+++ src/component/bookpage/TimeTable.module.css
@@ -31,3 +31,7 @@
     cursor: not-allowed;
   }
   
+  .timetable {
+    padding: 10px 0px 10px 15px;
+    border: 1px solid #D9DDDC;
+}
