--- src/component/bookpage/Times.module.css
@@ -1,4 +1,0 @@
-.timetable {
-    padding: 10px 0px 10px 15px;
-    border: 1px solid #D9DDDC;
-}
